Crate zstring

A crate for “thin pointer” zero-termianted string data.

Unlike the CStr and CString types, these types are compatible with direct FFI usage.

Structs

ArrayZString

An array holding textual data that’s zero termianted.

CharDecoder

Decodes byte sequences as if they were utf-8.

ZStr

Borrowed and non-null pointer to zero-terminated text data.

ZString

Owning and non-null pointer to zero-terminated textual data.

Enums

ZStringError

An error occurred while trying to make a ZStr or ZString.

Functions

zstrings_as_zstrs

Re-view a slice of ZString as a slice of ZStr