zstring

Struct ZStr

pub struct ZStr<'a> { /* private fields */ }

Borrowed and non-null pointer to zero-terminated text data.

Because this is a thin pointer it’s suitable for direct FFI usage.

The bytes pointed to should be utf-8 encoded, but the CharDecoder used to convert the bytes to char values is safe to use even when the bytes are not proper utf-8.

Safety

This is repr(transparent) over a NonNull<u8>.

The wrapped pointer points at a sequence of valid-to-read non-zero byte values followed by at least one zero byte.

When you create a ZStr<'a> value the pointer must be valid for at least as long as the lifetime 'a.

Implementations

impl<'a> ZStr<'a>

pub const fn from_lit(s: &'static str) -> ZStr<'static>

Makes a ZStr<'static> from a &'static str

This is intended for use with string litearls, but if you leak a runtime string into a static string I guess that works too.

const FOO: ZStr<'static> = ZStr::from_lit("foo\0");

Panics

If try_from would return an error, this will panic instead. Because this is intended for compile time constants, the panic will “just” trigger a build error.

pub fn bytes(self) -> impl Iterator<Item = u8> + 'a

An iterator over the bytes of this ZStr.

This iterator excludes the terminating 0 byte.

pub fn chars(self) -> impl Iterator<Item = char> + 'a

An iterator over the decoded char values of this ZStr.

pub const fn as_ptr(self) -> *const u8

Gets the raw pointer to this data.